feat: Support streaming
Unfortunately, I was not able to test it properly. I ran out of quota
after running the streaming twice. But it did work.

o1 did a decent job again with a pretty simple and straightforward
prompt. But I had to reiterate a couple of times to make the code
consistent: I ended up dumping on it my current JS files and asking it
to update it. This did the job. A couple of manual but really minor
changes and voila! It works.

The initial prompt was:

```
Update the code to use audio streaming: https://api.elevenlabs.io/v1/text-to-speech/:voice_id/stream
```